Standardize hooks being separate from providers.
