+++ app/src/components/app/songs/song-preview.tsx
@@ -6,7 +6,7 @@ import { Button } from "@/components/ui/button";
 import { PreviewWindow } from "@/components/controller/preview-window";
 import { Controls } from "@/components/controller/controls";
 import { useController } from "@/hooks/useController";
-import { useServices } from "@/hooks/services.provider";
+import { useServices } from "@/hooks/useServices";
 
 interface SongPreviewProps {
   getSong: () => ISong;
