+++ app/src/hooks/useServices.ts
@@ -1,0 +1,12 @@
+import { useContext } from "react";
+import { ServicesContext } from "./services.provider";
+
+export const useServices = () => {
+  const context = useContext(ServicesContext)
+
+  if (context === undefined)
+    throw new Error("useBroadcast must be used within a BroadcastProvider")
+
+  return context
+}
+
